Condensed files to navigate more easily. Also added a server for eventual deployment and began coding for an api.
