--- march-madness-draft/src/components/layout/BracketSubhead.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function BracketSubhead(){
-    return(
-        <header>
-            <h3>The Teams. The Games. The Bracket.</h3>
-        </header>
-    )
-}
-
-
-export default BracketSubhead;
