--- march-madness-draft/src/components/layout/Rules.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Rules(){
-    return(
-        <p>Rules Div</p>
-
-    )
-}
-
-
-export default Rules;
