--- march-madness-draft/src/components/layout/Teams.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Teams(){
-    return(
-       <p>Teams Div</p>
-    )
-}
-
-
-export default Teams;
