+++ march-madness-draft/src/components/pages/Nav.js
@@ -3,11 +3,11 @@ import { Link } from 'react-router-dom';
 
 
 
-function Header(count){
+function Nav(count){
     return(
         <header style={headerStyle}>
             <h1>Fantasy Basketball</h1>
-            <Link style={linkStyle} to="/">Rules  </Link> | <Link style={linkStyle} to="/Bracket">Bracket</Link> | <Link style={linkStyle} to="/Teams">Teams</Link> | <Link style={linkStyle} to="/Play">Play</Link>
+            <Link style={linkStyle} to="/">Rules  </Link> | <Link style={linkStyle} to="/Bracket">Bracket</Link> | <Link style={linkStyle} to="/Teams">Teams</Link> | <Link style={linkStyle} to="/Draft">Draft</Link>
         </header>
     )
 }
@@ -23,4 +23,4 @@ const linkStyle={
     textDecoration: 'none'
 }
 
-export default Header;
+export default Nav;
