+++ march-madness-draft/src/components/utils/API.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+export default {
+    getUsers: function() {
+        return axios.get("/api/users");
+    }
+}
